+++ src/main/java/com/example/datn/repository/ThuocTinhRepo/KieuQuatRepo.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 @Repository
 
-public interface KieuQuatRepo extends JpaRepository<KieuQuat, Long> {
+public interface KieuQuatRepo extends JpaRepository<KieuQuat, Integer> {
     @Query("SELECT kq FROM KieuQuat kq " +
             "WHERE (LOWER(kq.ten) LIKE LOWER(CONCAT('%', :query, '%')) " +
             "AND (:trangThai IS NULL OR kq.trangThai = :trangThai))")
